requirements.txt
appdirs==1.4.0
decorator==4.0.11
Django==1.10.5
django-extensions==1.7.6
django-user-agents==0.3.0
djangorestframework==3.5.3
ipdb==0.10.2
ipython==5.1.0
ipython-genutils==0.1.0
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
PyYAML==3.12
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
ua-parser==0.7.2
user-agents==1.0.1
wcwidth==0.1.7
